requirements.txt
celery==5.3.1
django==4.2.2
djangorestframework==3.14.0
django-celery-results==2.5.1
psycopg2-binary==2.9.6